Updates to dashboard, login, menu, redux, users, and migration for inventories, and create new files for add inventory functionality

+++ server/migrations/20200326145018-create-inventories.js
@@ -38,11 +38,11 @@ module.exports = {
             },
             price: {
                 allowNull: false,
-                type: Sequelize.DECIMAL
+                type: Sequelize.FLOAT
             },
             best_price: {
                 allowNull: true,
-                type: Sequelize.DECIMAL
+                type: Sequelize.FLOAT
             },
             last_purchase_dt: {
                 allowNull: true,
